pyproject.toml
[tool.poetry]
name = "simpan"
version = "0.1.0"
description = "Workflow automation platform"
authors = ["Nikhil Bola Kamath"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
pydantic = "^2.8.2"
celery = "^5.4.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
langgraph = "^0.2.18"
langchain-community = "^0.2.16"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
